Multicast damage animation implemented
The Multicast_AnimateDamage function is fully implemented and functional. When a "deal damage" notify is called by the server when playing an attack animation, a multicast function is called to provide visual feedback in sync.

Current Bug: When passing the attack info from the piece to the player pawn, the "Attacker" reference gets lost, becoming a null pointer. This might be because the defending piece is trying to pass its attack info, which would be empty.