Piece ID system debugged
The piece identification system has been debugged and now works as intended.

Pieces' unique IDs were not being read properly when using virtual functions to return each piece's ID. I replaced those functions with a single UPROPERTY in the parent piece class that is set in the editor of each child piece class.

Piece IDs are now properly read and used in the piece data table. The correct, corresponding information is now retrieved and use in the piece information pop-up widget, the attack confirmation widget, and the attack sequence graphic.