Camera interpolation crash fixed
Fixed the bug that was causing the game to crash when trying to interpolate the camera. I passed the piece references to the client function, instead of trying to retrieve them from the server inside of the client function.

Note: I've left all deprecated code from the previous C++-only version of the attack sequence in case there are any references to it I haven't noticed, and to use as a reference for recreating the sequence in script. I'll most likely remove the unneeded code and clean up the source once I finish the attack sequence.