Attack reaction implemented, server-side attack resolution mostly implemented
Fully-implemented the attack reaction, animating and highlighting both pieces and playing a corresponding animation on the attack graphic.

Implemented most of the server's attack resolution, by cleaning up and destroying both pieces when needed.

The server still needs to make the attacker still needs to move to the defender's old tile when necessary, or walk back to its original tile if it had needed to move closer.

Client-side attack resolution still needs to be implemented.