Fixed bug causing the piece info widget to appear instead of the attack confirmation
Fixed a bug that caused a piece's info widget to pop up instead of being selected as the target of an attack.

This was caused by the player state not updating before the player selected the target, making the client think that they were just selecting a piece.

I fixed this by instantly updating the client's player status before sending the RPC to the server, so that the player doesn't have to wait for the server to update their status.