FIxed bug where the same piece could not be selected after its info widget was cleared
Fixed a bug preventing the player from selecting a piece they had selected before that piece's info widget was removed.

This was caused because the game thought that the widget was already displayed, and since it had the same information, it didn't play an opening animation. I fixed this by making the widget play the opening animation if its render opacity is less than 1.0.