Created pyromancer ability confirmation widget and refactored attack sequence's camera movement
Created the pyromancer's active ability ("Fireball") confirmation widget, which is a child of the attack confirmation widget, since the ability is an altered form of the attack sequence.

I also refactored the way that the player's camera is interpolated in order to re-use the interpolation code modularly. But this means that the sequence continues as soon as the camera interpolation timeline starts, instead of waiting until it finishes, since timers aren't replicated.