Added particle effects to, tweaked, and debugged "Fireball" ability
Added particle effects to the power-up phase of the Pyromancer's "Fireball" ability.

Debugged the camera returning to the wrong position after the ability ended. This was caused by both the ability code and the attack code trying to move the camera. I added a parameter to control whether or not the attack sequence moves the camera when called.

Debugged piece highlights being called multiple times and at the incorrect times. This was caused by piece highlight flash functions being called again before the previous one finished. I fixed this by adding a "flash duration" variable and shortening the duration of each piece highlight flash from 1.0 seconds to 0.5.