Fixed position and alignment of ability info pop-ups
Ability info pop-ups now pop up anchored at the center of the ability icon. If the piece info widget is on the left side, the pop-up will be aligned and justified to the left. If the piece info widget is on the right side, it will be aligned and justified to the right.