Implemented backend functionality of temporary modifiers
Created a temporary modifier system that keeps a collection of currently applied modifiers for each piece, which change their stats.

Removed the CurrentStrength and CurrentArmor mutators so that strength and armor can only be changed by applying modifiers.

Added a new OnActiveEffectEnded function that is called when a temporary modifier caused by an active ability ends.

Added, but have yet to implemented, OnRep functions for CurrentStrength and CurrentArmor. These will instantly update the displayed strength and armor of both piece information widgets, if the piece whose strength or armor was changed is currently displayed by one of them.