Added color indicators for modified values
Strength and armor values in piece info widgets will turn green, red, or yellow depending on modifiers affecting the stat.

Text appears green if modifiers are adding a net positive value to the stat.

Text appears red if modifiers are adding a net negative value to the stat.

Text appears yellow if the stat has modifiers but they all cancel each other out, adding a net 0 value to the stat.

Text only appears white when no modifiers are applied to the stat.