Began implementing stat modifier pop-ups
Created Modifier and Modifier List widgets and implemented their update system.

When the strength stat is hovered over, a modifier list widget is successfully created and populated with modifier widgets.

Widget positions, layering, and destruction systems still need to be implemented and debugged. The system also needs to be copied to implement armor modifiers.