Added canvas slot to modifier list to manipulate its position
Added a canvas slot to the modifier list widget and an overlay to access it. The slot allows the widget to be moved without affecting the parent widget that it's attached to.

Also fiddled with some widget positioning, cleaned up the included headers in the player pawn class, and swapped the "cancel" and "confirm" buttons for the attack confirmation widgets.