Reducing feedback noise: attack sequence highlights
Reworked the initial white flashes that highlight both pieces before an attack sequence to be much shorter and less distracting.

Removed piece highlights that flashed to show the outcome of an attack. The character animations and attack graphic widget should be enough.

This is the final stable build with full-info tile highlights. From now on, tile highlights will be used sparingly, and won't be used to communicate piece occupation.