Added checkered tiles to board
The board is now checkered, with tiles alternating between a light and a dark texture beneath their highlight texture.

This makes the board easier to read and is a compromise between having piece alignment highlights and no highlights at all. It's easy to understand what's happening but isn't overwhelming with information.