WIP piece hovering and tile rework debugging
Attempted to bind an OnBeginCursorOver function to the piece mesh, which called the occupied tile's OnBeginCursorOver function. Attempting to bind this event to skeletal meshes has failed to work, and binding it to other components, like collision capsules, has produced strange results.

Also fixed a bug that prevented the player from seeing a reticle while waiting for their turn.