Note for fixing OnBeginCursorOver function
The OnBeginCursorOver function was not working for pieces because of a few reasons.

First, the parent skeletal mesh component needs to have a skeletal mesh that aligns with the piece's modular components. Instead of binding OnBeginCursorOver to every single modular component, it can just be bound to the parent component, which is an outline of the character. However, this means that components that aren't in that skeletal mesh won't be detected when hovered over (e.g. weapons).

Second, the skeletal mesh that the parent skeletal mesh component is using has to have a physics asset applied to it.

Lastly, the closest components to the player's mouse will block cursor events for components behind them, if that component blocks the "visibility" collision channel. The collision capsule component enclosed the rest of the piece's components within it, and it blocked the visibility channel. So when it was hovered over, the game would register that the cursor was over the capsule, and ignored anything behind it. By forcing the capsule to ignore the visibility channel, there is nothing blocking the cursor from hovering over the skeletal mesh component.

Lastly, remember that collision is disabled on the modular components. These cover the parent skeletal mesh and would block the cursor event like the capsule component if their collision was enabled and they blocked the visibility channel.

The actual change I made in this commit is irrelevant. I just wanted to explain this feature.