Implemented piece selection reticle system
Pieces' tiles now gain a green reticle when they are selected by the player to perform an action (i.e. when a friendly piece is selected to bring up its piece information widget). The green reticle is not removed via hovering so that it remains as long as the piece is selected. It is removed manually when the piece is no longer selected.

This is done with the bReticleControlledByCursor variable, which determines whether or not cursor events (i.e. hovering and unhovering) will affect a tile's reticle.

This also required me to set the player's state to "SelectingAction" when they have a piece information widget open, which was already supposed to happen (I'm not sure how I managed to skip over the implementation of an entire player state). The player whose turn it is enters the SelectingAction state when opening a piece info widget.

This also works when the player is waiting for their turn, allowing them to also select pieces to perform actions with, in preparation for their turn.

Also reworked some piece placement code to make it more efficient. Additionally, the ready button during the piece placement phase becomes disabled when clicked.