Fixed a bug where tile's occupying pieces wouldn't change
This happened because I had commented out the entirety of the SetOccupyingPiece function instead of just the OnRep call when testing if I actually needed the OnRep for anything.

Also removed resetting the selected piece as soon as the attack confirmation widget was created. I like leaving the piece selected while confirming the attack, but it might cause some unexepected problems with selections during confirmations. I've removed it for now, but may add it back later.