Implemented blocking checks
Implemented checks to see if a target tile or piece for a movement or ability is blocked by another piece. This just checks every tile between the current and target ones with a line trace. If any tiles between them are occupied, then that piece can't be reached. Perfectly diagonal tiles (e.g. 1 forward, 1 left) are valid, but I might change that later depending on how the game plays.

Also made piece and tile brightness a parameter that can be changed in the editor, because the original values I was testing with were way too bright for darker environments.