Reworked ClearSelection system and allowed players to change between moving and attacking
Reworked the ClearSelection system of the player pawn so that each selected actor (the selected piece, selected enemy piece, selected target piece, and selected tile) can be reset individually.

Also implemented a mechanic that allows players to switch their targets of move actions between empty and enemy-occupied tiles. Before, this would cause problems because the confirmation widgets of a move and of an attack are different. Now, if the widget type needs to be switched, it will destroy the previous widget.