Fixed ensure problem caused when creating "Fireball" confirmation widget
Fixed an ensure problem caused when creating the Pyromancer's active ability confirmation widget. This happened because I was calling the super native constructor, which binds the "attack" and "cancel" buttons to the attack confirmation functions. When I tried to re-bind the buttons to the Pyro's active ability functions, it caused problems with the delegates.

I fixed this by moving the binding logic into a BindButtons function that can be overridden by each child class to bind their own functions to each button. The attack confirmation widget's NativeConstruct calls the correct class's BindButtons function when the widget is created.