Added Academy Recruit classes and added limited uses to "Fireball" ability
Added the C++ and blueprint classes for the Academy Recruit piece to the game, and began designing its appearance.

The Pyromancer's "Fireball" ability is really strong, being able to attack with 14 strength without any risk. I added a limited number (3) of uses to it to balance it a bit. I'm expecting it to still be very powerful, so I might nerf the Pyromancer's base strength in the future.

Implementing the first ability with limited uses revealed a few bugs with the limited-use ability system, which I fixed.