Fixed Academy Recruit's animations
The Academy Recruit piece was not being animated. This was because the parent skeletal mesh component had its "Visibility Based Anim Tick Option" variable set to only animate when it was being rendered. This meant that when the parent mesh was invisible, nothing was animating.

I fixed this by setting the variable in the AParentPiece class to always animate, regardless of whether or not the mesh is visible.

Also began further implementing the Academy Recruit piece and standardizing various settings across pieces.