Began implementing Recruit's passive ability
Began implementing the Recruit piece's passive ability, "Soldier's Bond." This requires a new OnMoveToTileCompleted function for pieces and a complete rework of the modifier system.

I finished the skeleton code of the passive ability, which triggers when a Recruit piece finishes moving to a new tile. Recruits now gain a modifier when they finish moving depending on how many Recruits they are adjacent to on their new tile.