Moved the AddModifier function to the player pawn from the parent piece
Moved the AddModifier function to the player pawn class. This allows the function to be networked, so that modifiers can be added by any client.

The player pawn handles adding the actual modifier to the piece and any visual feedback that needs to be displayed. The piece handles changing its own value based on the new modifier with an OnRep function.

This structure allows pieces' strength and armor variables to stay protected. They can only be changed by adding and removing modifiers. They cannot be edited directly.