Debugged RemoveModifier() and reworked StartGame()
Debugged the re-implementation of the RemoveModifier function, specifically its use in recruits. The problem wasn't with the function, but rather clients failing to recognize when a pawn is locally controlled.

I'm honestly not sure why this wasn't working. So instead of fixing the GetAlignment function, I replaced it by instead checking if the target recruit has the same instigator as the recruit being tested, which would mean that they are both controlled by the same player.

I also reworked how the game starts. Instead of trying to start the game every time a player's bReady variable changes, and then checking within the StartGame function if the function is being called by the server, I instead made the StartGame function a server RPC that gets called by whichever player is the last to click "Ready."