Implemented queued modifier pop-ups
Modifier pop-ups are now triggered when adding or removing modifiers. If a piece is already playing a pop-up animation, the next pop-up will be "queued." Pop-ups are automatically triggered sequentially until every modifier has played any necessary pop-ups.

Also implemented dual modifiers, which modify both a piece's armor and strength statistic. When these are added, both a strength pop-up and an armor pop-up will be queued, so that there's one for each stat, which play one-after-the-other.

I also debugged a problem that prevented clients from seeing the other player's pieces when the game started if it was started by a listen server. This happened because the function that reveals pieces was a server RPC, so pieces weren't revealed for clients.