Debugged Recruit's passive ability on game start
There was a lot of unexpected behaviors and bugs when recruits were placed adjacently at the start of the game. It turns out that these were all caused by having the OnGameStart function call to a multicast function in the game state. This caused Recruits to try to add modifiers for their passive ability multiple times when the game started.

I also added a check to ensure that Recruits don't try to modify other Recruits' modifiers when the game starts, only their own.