Set up the King piece's blueprint defaults and mesh
Set up the King piece's default blueprint settings.

Implemented the King's mesh, importing the "Polgyon Fantasy Characters" asset pack to use their premade King character.

This is the first piece to use a premade mesh in place of a modular character. I created a new material for it, which has the "Brightness" parameter, "Fresnel" parameter, and emissive color nodes. This allows the fresnel and highlight code to work with the King's mesh, even though its mesh's material is different than the other pieces' materials'.