Implemented the King piece's base animations
Fully implemented the King piece's idling, movement, and combat animations, allowing it to successfully participate in fights, either as an attack or as a defender.

The King is the first piece that is not modularly designed, which means that it is also the first piece not to use the modular character skeleton. With some retargeting and some minor blueprint tweaks, I was able to seamlessly insert this skeleton into the game's animation system. This means that the prebuilt animation system is compatible with any skeleton.

The King is also the first piece to use any animations that are not exported from Mixamo--i.e. animations that don't use Mixamo's skeleton. These just had to be retargeted to the King's skeletal mesh's skeleton and they were able to be use in the animation system.