Implemented unique damage effects
Changed the "damaged effect" variable to "damaging effect." Now, the damage effect played on a piece when it takes damage is determined by the piece attacking it. This makes more sense, since the type of attack should dictate how a piece appears to take damage, rather than the piece itself.