Finished parent commander class and reworked modifier pop-ups
Finished the parent "commander" class. When a piece of the "commander" class dies, it triggers the game state's EndGame() function, passing in the correct winner.

Reworked the modifier pop-up class. Modifiers are now attached to the "pop-up location component" of the actor that spawned them, if that actor has the component and the pop-up was requested to do so. This means that pop-ups will follow piece's animations and movements and that this can be modularly expanded to use modifier pop-ups on other types of actors.

Also reworked how the pop-up is animated and updated. The new system uses a timeline that interpolates between the desired values for height and scale rather than interpolating between 0.0 and 1.0 and using arithmetic to figure out the correct transform values to use. This new method is way simpler and much easier to adjust and expand.

Also tweaked the pop-up animation to improve the feel and smoothness of it.