Refactored attack confirmation system
Slightly refactored the attack confirmation system (and the move confirmation system) to allow players to switch between move and attack actions depending on the tile they select, without having to cancel their action and start over.

This required a lot of minor changes and reworks, which did end up goofing with the active ability confirmation system. Fortunately, that was already the next system I was planning on reworking, so hopefully everything will be smoothed out in the next patch.