Refactored Pyromancer and King's active ability confirmation sequences
Refactored the Pyromancer and King's active ability confirmation sequences and widgets. Both abilities now work with the previously reworked active ability confirmation system.

Also had to rework the client-side prediction code for the parent piece class because of how authority works there. In the player state, we can always make our instant update locally and then update the server if we don't have authority over our player state, which happens whenever we are playing on a non-server client. But the parent piece does not have an owning connection, so no one has authority and a server RPC is always needed to update it. I don't understand this at all, but it works now so we'll just roll with it.