Implemented ability range previewing
Implemented the option to preview abilities without entering a confirmation state. When abilities are hovered over in the piece info widget, all valid targets will be highlighted.

This allows players to quickly and easily preview actions. More importantly, this allows players to preview abilities when it's not their turn or when the piece's ability is on a cooldown. Since the "use active" button is hidden or disabled in these cases, previewing the ability would be otherwise impossible.

To do this, I created and implemented a new TargetInterface class with a HighlightTarget pure virtual function. Each class that can be targeted by an ability implements this interface to highlight themselves.

Currently, targets remain highlighted even when the player stops hovering over the ability box. A new RemoveTargetHighlight pure virtual function will be implemented next to fix this.