Adjusted the positioning of the modifier list pop-up
When viewing a piece with active modifiers, the modifier pop-up's position is determined by the affected stats.

If the piece's modifiers are only affecting its strength stat, or if the piece's modifiers are affecting both its strength and armor stats, then the pop-up is positioned at the top of the stat box, above the piece's displayed strength.

If the piece's modifiers are only affecting its armor stat, then the pop-up is positioned lower, above the piece's displayed armor.