Implemented action state checks in piece info widget
When the player tries to use a movement or ability action from the piece info widget, it will check their current player status. If they are not in the "selecting action" state, their input will not be registered and the widget will be refreshed.

This should never happen, but it acts as another safeguard against latency issues and hacking.