Added a reticle to selected enemy pieces
A reticle will now indicate which enemy piece a player is currently viewing the information of. Reticles were already used to indicate targets of attacks or abilities, but there were not reticles while selecting and viewing pieces. Reticles are now used in the same way for both friendly and enemy pieces.