Added individual stat buttons
There are now 3 buttons that can be used to hover over modified statistics.

If only a piece's strength is modified, the player can only hover over that piece's strength stat in the piece info widget to view its temporary modifiers.

If only a piece's armor is modified, the player can only hover over that piece's armor stat.

If both of a piece's stats are modified, the player can hover anywhere over their stats.