Fixed unhighlighting after confirmation bug
Fixed a bug that prevented selected pieces from being unhighlighted after confirming an action that targeted them.

This happened for two reasons. One was that while trying to revert back to the default highlight, a different highlight would be activated (like the white flash that indicates attacks), which interrupted the original one. I fixed this by skipping to the end of the highlight animation whenever an active highlight is interrupted.

The other problem was a small bug with the Pyromancer's overridden DestroyWidget function, which treated the selected enemy piece as the target of the ability instead of the selected target piece. I'm not worried about this, since the Pyromancer's confirmation widget is a very special case.