Implemented custom information for attack preview widgets
Custom information can now be displayed in attack preview/confirmation widgets, including an override of the pieces' displayed stats. This was designed for the pyromancer's active ability, which required the preview to treat the pyromancer as if it had +X strength, because of its ability's buff. This will also be used for future pieces.

This new method of passing in a pre-built structure of information regarding the attack preview also allows for different information to be displayed depending on whether or not a one-sided attack is being previewed.