Added the removal of ability effects after a piece dies
When a piece dies, any actors attached to it (i.e. any effects) are also destroyed. This required some fancy footwork (specifically with the Guardian's shield) to implement destruction animations whenever an effect is requested to be destroyed.

I also reworked the Guardian's shield effect, giving it a pulsing effect instead of a curtain effect. The speed of this effect is randomized, making each shield look unique. I also added pop-up and destruction animations to the effect.