Fixed client-side temporary highlight bug
Fixed a bug that prevented pieces' highlights from returning to their original highlight when being highlighted temporarily.

When a highlight is interrupted with another highlight request, instead of manually trying to set the piece's highlight information to what it would be if the previous highlight had finished, the piece will now just skip to the end of its previous highlight before beginning the new one.