Began reworking camera and fixed Guardian rotation bug
Began reworking the camera movement system. Vertical movement works but horizontal movement is still bugged.

Fixed the bug that caused the Guardian to rotate improperly after using its ability on the client. This was because both the server and the client were requesting the Guardian reset its rotation, so it tried rotating twice simultaneously. Adding an authority check fixed this.