Began debugging piece drag system and implementing main menu
Began debugging the piece drag system. Cancelling a drag while dragging a piece breaks the game because the widget doesn't have enough time to spawn the piece. Unreal automatically destroys the dragged widget as soon as the drag is cancelled and this cannot be overridden with any method I've found. I'll continue working on this.

Began implementing a main menu by creating a new main menu map and a new "display piece" blueprint, which is a visual replica of another piece, but without any additional overhead code that would go unused in a menu.

Also fixed a little mistake where the knight piece's animation blend space was using the default idle/walk animations instead of the knight's idle/walk animations.