Implemented main menu navigation framework
Implemented the framework for the main menu navigation system. With this system, the player can use the HUD as an interface to switch between any widgets using a "queue" feature. When the player wants to change widgets, they "queue" their new widget in the HUD. The HUD will play its current widget's "deactivation" animation. When the animation finishes, the HUD automatically switches to the queued widget, playing its "activation" animation.

This method is fully modular, allowing each widget the player can interact with to be designed independently, and to use any kind of animations it wants without worrying about menu transitions. This is also a more cost-effective system than having one massive widget that just hides or reveals parts of each menu, or constantly have every widget open and just switching between which ones are visible and one screen.