Implemented in-menu matchmaking
Implemented in-menu matchmaking, allowing players to stay in the menu while searching for a match. If there are no available matches, the player will create a new session and reload their main menu map as a listen server. If available sessions were found, the player will connect to that session, the host will travel to a new game map, and the game will start.