Implemented menu queue illusion and matchmaking start pop-up
Added some black magic trickery that makes it look like players are not changing levels when queueing for a match. When the player loads into the main menu level, it will check if they are currently hosting a session. If they are, it moves them back into the play menu and creates a queue timer, so it looks like they never left the play menu. This does cause players to freeze when they begin matchmaking, but this is the best we can do without dedicated servers.

Also implemented a new matchmaking start pop-up, which displays "starting matchmaking..." when the player clicks "quickplay." This acts as a buffer between the player requesting to search for a session and creating a new session. In other words, "starting matchmaking" displays while searching for a session or creating a new one, whereas the queue timer widget displays while waiting for someone else to join a session you're hosting.

Lastly, for some reason I had written some UI management code in the main menu player controller. I moved this code to the main menu HUD.