Tweaked the in-menu session creation
Tweaked the way sessions are created and traveled to as the host. While searching for sessions, navigation buttons are now disabled so the player is forced to temporarily remain in the play menu. The play menu will also now only play an activation animation if they are NOT in a session when loading in. This is to add to the illusion that the player isn't actually changing levels.