Implemented map selection menu
Implemented the new map selection menu. Players can enter the menu from the lobby menu to change the selected map. Each map in the "playable maps" map pool will be listed. Players can click on each map listing to display its information within the map selection menu. Selecting the new map and actually updating the match settings are not yet implemented.

Also had to fix another missing enumerator bug, which caused the EAttackResult enum to be missing its last member. This also happened with the EAlignment enum a long time ago. I fixed both by adding another member to the enum (a "debug attack/alignment" option), which replaces the member that was previously last as the missing member. It's dumb but it works.