Implemented saved custom game settings
Implemented a new game state class for the main menu that contains a structure of the settings for the current custom game, if the local player is in a custom lobby. When selecting a map in the map selection menu, it will be updated in the game state. The lobby menu will then update its displayed selected map by retrieving the game state's data.