Added lobby exit confirmation and lobby menu navigation on connection
Players are now given a confirmation pop-up when they try to leave a lobby. When confirmed, they will return to the play menu. The destruction of the lobby has not been implemented yet.

Also implemented lobby navigation on connection success. When players successfully connect to a custom lobby via their beacon client, they will navigate to the lobby menu.