Implemented lobby disconnect delegate
Implemented the OnDIsconnect delegate in the custom lobby. When a client is disconnected by a host (like when the host quits the lobby), the client is notified and can perform their own client-side logic. Now, when the client is disconnected, they will be returned to the play menu, instead of staying in the lobby menu.