Implemented replicated lobby settings and information
The lobby's information (i.e. settings) is now stored on the host beacon instead of in the game state. The lobby settings are replicated to connected clients. When a client connects, the host will send the lobby's settings at the time of connection for the client to replicate. Whenever the lobby's settings are updated, all connected clients are notified of the changes and replicate them.