Implemented lobby friend list and created invitation widgets
Implemented the lobby friend list widget. When the player clicks "invite friends" while in a custom lobby menu, a list of their friends will appear, displaying all of their friends' names, avatars, and statuses (offline, in-game, etc.).

Also created the invitation widget that will pop-up when receiving a lobby invitation in the future.