Finished implementing Assassin piece
Finished fully implementing the Assassin piece. The piece's active ability is fully functional and its animations and particle effects are implemented. I also tweaked several smaller details (the materials, animations, etc.), refined the animations, added a displayed passive ability, and fixed some smaller bugs with the systems this piece uses.

This is the first piece with an asymmetrical movement pattern or ability range, meaning I needed to change how tiles are compared. For example, "backwards 1 tile" for one player is actually forwards for the other, which was a massive problem with the system.

To fix this, I implemented a new CheckTilePosition function to the ABoardTile class. This function takes in another board tile and a pair of coordinates. It will check the direction that its occupying piece is facing and compare it to its own direction (all tiles always face the same direction) using a dot product. If the piece is facing the opposite way of the tile, then the coordinates are applied in reverse.

This new coordinate comparison method works perfectly with the Assassin's active ability. Next, I will update all other pieces' movement pattern and range functions to use it too.