Fixed dragging piece card text bug
Fixed a bug that was causing pieces' name text to remain briefly on the screen when their card was dragged during the piece setup.

I don't think this was actually a bug, but rather just some goofiness with animating nested widgets. I fixed this by animating the actual piece widget instead of animating the drag widget, which contained the piece widget.