Implemented visual effects and debugged passive ability for War Mage piece
Implemented the visual effects for the War Mage, including a custom attack particle effect.

Also debugged the War Mage's passive ability. This required reworking a good bit of the modifier system, but it is now more dynamic, as you can now control which modifiers stack and when, and each modifier now has a unique ID to different modifiers of the same type.