Reworked lobby friends list to only display online friends
Pretty much gave up on trying to get Steam app names from the Steamworks API. Now, only friends who are currently playing Kingdoms will be displayed in the friends list when in lobbies.

I'm frankly not sure what the other 400 changed files are from. It seems like the RiderLink plugin had to be rebuilt, though I'm not sure why we're tracking that with Git.