Implemented start button in custom lobbies
Implemented a start button in custom lobbies. Hosts now have a "start" button in their lobby menu that is disabled until two players are present in the lobby. Once two players are in the lobby, the button can be pressed. The host will create a new game server and travel to the currently selected map as a listen server. All clients will then follow the host to the map and the game will begin.