Implemented joining lobbies from launch and added dynamics to "invite friends" button
Implemented the ability to join lobbies from launch. When players accept invitations or request to join friends' lobbies, but they don't have the game open, the game will automatically launch with a command that automatically joins the lobby.

The "invite friends" widget in the lobby menu now dynamically enables/disables depending on if the lobby is full or not.